requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
et-xmlfile==1.1.0
idna==3.2
numpy==1.21.2
openai==0.10.4
openpyxl==3.0.7
pandas==1.3.3
pandas-stubs==1.2.0.17
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.6
